Refactor evaluator into modules and add tick helpers
